requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
selenium>=4.18
scikit-learn>=1.3
tensorflow>=2.15
pulp>=2.7
lxml>=4.9
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0
